small modifications to the unit tests for classes RoutingTable and BestPheromonePolicy. The unit tests now create an instance of LinearEvaporationPolicy which in turn is passed to the constructor of the RoutingTable. This is required to do so in order to set the interval for the checkForEvaporation method so high that it does not decrease the pheromone value of the test routing table entries.

+++ include/Time.h
@@ -35,8 +35,8 @@ namespace ARA {
      */
     class Time {
         public:
-          float getTimeDifferenceInSeconds(struct timeval*, struct timeval*);
-          float getTimeDifferenceInMilliseconds(struct timeval*, struct timeval*);
+          int getTimeDifferenceInSeconds(struct timeval*, struct timeval*);
+          int getTimeDifferenceInMilliseconds(struct timeval*, struct timeval*);
 
         private:
           /// the method returns the time difference in seconds
